extend 'conda env create'  tests to 'conda create' too
